# unoq/schemas/openapi-v1.yaml
openapi: 3.1.0
info:
  title: Heimdall UNO Q API
  version: 1.0.0
servers:
  - url: /api/v1
paths:
...
  /clips:
    get:
      operationId: listClips
      responses:
        "200":
          description: Protected capture clips
    post:
      operationId: createClip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClip"
      responses:
        "202":
          description: Clip armed and collecting its post-trigger interval
  /clips/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      operationId: downloadClip
      responses:
        "200":
          description: Completed protected capture ZIP
          content:
            application/zip:
              schema:
                type: string
                format: binary
    delete:
      operationId: deleteClip
      responses:
        "204":
          description: Clip deleted
        "404":
          description: Clip not found
...
components:
  schemas:
    CreateClip:
      type: object
      properties:
        name:
          type: string
          maxLength: 120
        note:
          type: string
          maxLength: 2000
...